refactor: rename tdp operations to tdp ops
