+++ tdp/cli/__main__.py
@@ -13,7 +13,7 @@ from tdp.cli.commands.dag import dag
 from tdp.cli.commands.default_diff import default_diff
 from tdp.cli.commands.deploy import deploy
 from tdp.cli.commands.init import init
-from tdp.cli.commands.operations import ops
+from tdp.cli.commands.ops import ops
 from tdp.cli.commands.plan import plan
 from tdp.cli.commands.playbooks import playbooks
 from tdp.cli.commands.status import status
