+++ tdp/cli/commands/ops.py
